requirements.txt
blis==0.2.4
certifi==2019.9.11
chardet==3.0.4
ChatterBot==1.0.5
cymem==2.0.3
Django==2.2.7
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
idna==2.8
jsonfield==2.0.2
mathparse==0.1.2
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
Pint==0.9
plac==0.9.6
preshed==2.0.1
pymongo==3.9.0
python-dateutil==2.7.5
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.3.0
srsly==0.2.0
thinc==7.0.8
tqdm==4.38.0
urllib3==1.25.7
wasabi==0.4.0
